Add i18n for gift and leaderboard msgs
Summary:
Basically, for the two messages, we only need to add them to ios.json and android.json.
But the old clients only support accept the i18n messages instead of tokens.
So we need to translate the i18n at server side and that's why we add the messages to backend.json, too.

For now, we just send both messages and tokens to client

Test Plan: none

Reviewers: #backend, abner

Reviewed By: #backend, abner

Differential Revision: https://media17.phacility.com/D2766